# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/brands/all": {
      "get": {
        "operationId": "BrandsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/brands": {
      "post": {
        "operationId": "BrandsController_create",
        "parameters": [
          {
            "name": "createdb",
            "required": false,
            "in": "query",
            "description": "create database for brand",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/brands/{brandId}/account-manager": {
      "delete": {
        "operationId": "BrandsController_unassignAccountManager",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/brands/{brandId}": {
      "get": {
        "operationId": "BrandsController_getBrand",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "BrandsController_deleteBrand",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "BrandsController_update",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/brands/{brandId}/user/{userId}": {
      "get": {
        "operationId": "BrandsController_assignUser",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "BrandsController_removeUserFromBrand",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/brands/{brandId}/verify": {
      "get": {
        "operationId": "BrandsController_verify",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsersDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/all": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "account_type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsersDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/verify": {
      "post": {
        "operationId": "UsersController_verifyByAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/brands": {
      "get": {
        "operationId": "UsersController_getBrandsByUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/brands": {
      "get": {
        "operationId": "UsersController_getBrands",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/permissions": {
      "get": {
        "operationId": "UsersController_userPermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/permissions": {
      "get": {
        "operationId": "UsersController_Permissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/account-manager": {
      "put": {
        "operationId": "UsersController_assingAccountManager",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "brandsId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_unassignAccountManager",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "brandsId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Vendo API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateBrandDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "Pending",
              "Active",
              "Inactive"
            ],
            "description": "Status of the brand"
          },
          "name": {
            "type": "string",
            "description": "Name of the brand"
          },
          "usersId": {
            "type": "number",
            "description": "User ID"
          },
          "u_amazon_seller_name": {
            "type": "string",
            "description": "Amazon seller name"
          },
          "u_amazon_marketplace_name": {
            "type": "string",
            "description": "Amazon marketplace name"
          },
          "allowed_user_id": {
            "type": "string",
            "description": "Allowed user ID"
          },
          "u_cust_id": {
            "type": "string",
            "description": "Customer ID"
          },
          "u_card_last_no": {
            "type": "number",
            "description": "Card last number"
          },
          "access_token": {
            "type": "string",
            "description": "Access token"
          },
          "password_reset_token": {
            "type": "string",
            "description": "Password reset token"
          },
          "auth_key": {
            "type": "string",
            "description": "Auth key"
          },
          "u_refunded": {
            "type": "number",
            "description": "Refunded"
          },
          "u_canceled": {
            "type": "number",
            "description": "Canceled"
          },
          "u_payment": {
            "type": "number",
            "description": "Payment"
          },
          "u_mws_seller_id": {
            "type": "string",
            "description": "MWS seller ID"
          },
          "u_mws_auth_token": {
            "type": "string",
            "description": "MWS auth token"
          },
          "u_mws_status": {
            "type": "number",
            "description": "MWS status"
          },
          "u_is_logged": {
            "type": "number",
            "description": "Is logged"
          },
          "u_is_auto_logout": {
            "type": "number",
            "description": "Is auto logout"
          },
          "u_aff_username": {
            "type": "string",
            "description": "Affiliate username"
          },
          "u_aff_password": {
            "type": "string",
            "description": "Affiliate password"
          },
          "u_affiliate_id": {
            "type": "string",
            "description": "Affiliate ID"
          },
          "u_affiliate_status": {
            "type": "number",
            "description": "Affiliate status"
          },
          "u_sub_plan": {
            "type": "string",
            "description": "Subscription plan"
          },
          "u_refer_date": {
            "format": "date-time",
            "type": "string",
            "description": "Refer date"
          },
          "db_name": {
            "type": "string",
            "description": "Database name"
          },
          "u_db_server": {
            "type": "string",
            "description": "Database server"
          },
          "u_db_username": {
            "type": "string",
            "description": "Database username"
          },
          "u_db_password": {
            "type": "string",
            "description": "Database password"
          },
          "u_enterprise_id": {
            "type": "number",
            "description": "Enterprise ID"
          },
          "u_first_google_sheet_name": {
            "type": "string",
            "description": "First Google Sheet name"
          },
          "u_second_google_sheet_name": {
            "type": "string",
            "description": "Second Google Sheet name"
          },
          "u_first_google_sheet_id": {
            "type": "string",
            "description": "First Google Sheet ID"
          },
          "u_second_google_sheet_id": {
            "type": "string",
            "description": "Second Google Sheet ID"
          },
          "amazon_advertising_access_token": {
            "type": "string",
            "description": "Amazon advertising access token"
          },
          "amazon_advertising_refresh_token": {
            "type": "string",
            "description": "Amazon advertising refresh token"
          },
          "amazon_advertising_token_type": {
            "type": "string",
            "description": "Amazon advertising token type"
          },
          "amazon_advertising_access_token_expiry": {
            "type": "number",
            "description": "Amazon advertising access token expiry"
          },
          "amazon_advertising_profile_id": {
            "type": "string",
            "description": "Amazon advertising profile ID"
          },
          "amazon_advertising_marketplace_id": {
            "type": "string",
            "description": "Amazon advertising marketplace ID"
          },
          "amazon_advertising_client_id": {
            "type": "string",
            "description": "Amazon advertising client ID"
          },
          "amazon_advertising_client_secret": {
            "type": "string",
            "description": "Amazon advertising client secret"
          },
          "amazon_advertising_return_url": {
            "type": "string",
            "description": "Amazon advertising return URL"
          },
          "amazon_advertising_response": {
            "type": "string",
            "description": "Amazon advertising response"
          },
          "rds_credential_id": {
            "type": "number",
            "description": "RDS credential ID"
          },
          "server_credential_id": {
            "type": "number",
            "description": "Server credential ID"
          },
          "u_is_assigned": {
            "type": "number",
            "description": "Is assigned"
          },
          "u_active_asin_count": {
            "type": "number",
            "description": "Active ASIN count"
          },
          "u_allowed_report_count": {
            "type": "number",
            "description": "Allowed report count"
          },
          "pending_reports": {
            "type": "number",
            "description": "Pending reports"
          },
          "avg_order_per_day": {
            "type": "number",
            "description": "Average order per day"
          },
          "avg_sku": {
            "type": "number",
            "description": "Average SKU"
          },
          "business_report_is_done": {
            "type": "boolean",
            "description": "Business report is done"
          },
          "business_report_count": {
            "type": "number",
            "description": "Business report count"
          },
          "previous_business_report_dates": {
            "type": "string",
            "description": "Previous business report dates"
          },
          "created_at": {
            "type": "number",
            "description": "Created at"
          },
          "updated_at": {
            "type": "number",
            "description": "Updated at"
          },
          "created_by": {
            "type": "number",
            "description": "Created by"
          },
          "updated_by": {
            "type": "number",
            "description": "Updated by"
          },
          "last_update": {
            "type": "number",
            "description": "Last update"
          },
          "syncing_data": {
            "type": "string",
            "description": "Syncing data"
          },
          "sync_progress": {
            "type": "string",
            "description": "Sync progress"
          },
          "sync_last_update": {
            "type": "number",
            "description": "Sync last update"
          },
          "table_column_configuration": {
            "type": "string",
            "description": "Table column configuration"
          },
          "days_of_inventory": {
            "type": "number",
            "description": "Days of inventory"
          },
          "u_last_cron_time": {
            "type": "number",
            "description": "Last cron time"
          },
          "u_cron_status": {
            "type": "number",
            "description": "Cron status"
          },
          "u_applied_coupon_code": {
            "type": "string",
            "description": "Applied coupon code"
          },
          "all_report_requested": {
            "type": "boolean",
            "description": "All report requested"
          },
          "requested_report_types": {
            "type": "string",
            "description": "Requested report types"
          },
          "u_auto_sync": {
            "type": "number",
            "description": "Auto sync"
          },
          "report_id": {
            "type": "number",
            "description": "Report ID"
          },
          "end_date": {
            "type": "number",
            "description": "End date"
          },
          "start_date": {
            "type": "number",
            "description": "Start date"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}